plot3d::point_match

Function point_match

pub fn point_match(
    x: f64,
    y: f64,
    z: f64,
    x2: &[f64],
    y2: &[f64],
    z2: &[f64],
    nu: usize,
    _nv: usize,
    tol: f64,
) -> Option<(usize, usize)>

Find the indices of the closest point in a 2D face grid to (x, y, z).

The face grid is represented by flat arrays x2, y2, z2 of length nu * nv, stored row-major (u varies fastest).

Returns Some((u, v)) if the minimum distance is within tol, otherwise None.